refactored to use methods when involving type structs !Untested!

+++ models/NutritionalValueModel.go
@@ -1,19 +1,19 @@
 package models
 
 type NutritionalValue struct {
-	ID		  		uint 	`gorm:"primarykey"`
-	OwnerTitle		string 	`json:"owner_title"`
+	ID		  		uint 		`gorm:"primarykey"`
+	OwnerTitle		string 		`json:"owner_title"`
 	OwnerID   		uint
   	OwnerType 		string
-	Edited 			bool   	`json:"edited" gorm:"ignore"`
+	Edited 			bool   		`json:"edited" gorm:"ignore"`
 
-	Kcal			int    	`json:"kcal"`
-	Kj				int	   	`json:"kj"`
+	Kcal			float64    	`json:"kcal"`
+	Kj				float64	   	`json:"kj"`
 
-	Fat				int    	`json:"fat"`
-	SaturatedFat	int	   	`json:"saturated_fat"`
-	Carbohydrate	int	   	`json:"carbohydrate"`
-	Sugar			int    	`json:"sugar"`
-	Protein			int    	`json:"protein"`
-	Salt			int	   	`json:"salt"`
+	Fat				float64    	`json:"fat"`
+	SaturatedFat	float64	   	`json:"saturated_fat"`
+	Carbohydrate	float64	   	`json:"carbohydrate"`
+	Sugar			float64    	`json:"sugar"`
+	Protein			float64    	`json:"protein"`
+	Salt			float64	   	`json:"salt"`
 }
